+++ R/PlotColours.R
@@ -177,57 +177,65 @@
   )
 
   if (palette_type == "qualitative") {
-    # Default qualitative palettes are based on the Tableau 10 palette by
-    # Maureen Stone, Cristy Miller and Jeffrey Heer.
+    # Palette derived from:
     #
-    # * Maureen Stone, Designing Colors for Data, International Symposium on
-    # Computational Aesthetics in Graphics, Visualization, and Imaging, Banff,
-    # AB, Canada, June 22, 2007.
+    # colorspace::diverge_hcl(n = 21, h = c(245, 35), c = c(130, 85), l = c(65,
+    # 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
     #
-    # * Jeffrey Heer and Maureen Stone, Color Naming Models for Color Selection,
-    # Image Editing and Palette Design, ACM Human Factors in Computing Systems,
-    # 2012.
+    # colorspace::diverge_hcl(n = 21, h = c(185, 10), c = c(130, 85), l = c(60,
+    # 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
     #
-    # * https://www.tableau.com/about/blog/2016/7/colors-upgrade-tableau-10-56782
-    # 
+    # colorspace::diverge_hcl(n = 21, h = c(65, 125), c = c(100, 85), l = c(70,
+    # 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
+    #
+    # colorspace::diverge_hcl(n = 21, h = c(210, 320), c = c(60, 85), l = c(60,
+    # 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
+    #
+    # colorspace::diverge_hcl(n = 21, h = c(100, 0), c = c(50, 85), l = c(70,
+    # 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
+    
+    palette_10 <- c(
+      "#00A8FF",
+      "#F77D00",
+      "#00BCAB",
+      "#FE535E",
+      "#D4A600",
+      "#40C41C",
+      "#00A2B7",
+      "#C973B6",
+      "#9FB368",
+      "#E495A5"
+    )
+    
+    palette_20 <- c(
+      "#00A8FF",
+      "#93BDF0",
+      "#F77D00",
+      "#E9AD8D",
+      "#00BCAB",
+      "#96C9C5",
+      "#FE535E",
+      "#DFB6B7",
+      "#D4A600",
+      "#CEC2AC",
+      "#40C41C",
+      "#B2C9B0",
+      "#00A2B7",
+      "#8AC4D1",
+      "#C973B6",
+      "#DCADD0",
+      "#9FB368",
+      "#B3C680",
+      "#E495A5",
+      "#ECB1BC"
+    )
+    
     if (!use_alternative) {
       if (n <= 10L) {
-        colours <- c(
-          "#4e79a7",
-          "#f28e2b",
-          "#e15759",
-          "#76b7b2",
-          "#59a14f",
-          "#edc948",
-          "#b07aa1",
-          "#ff9da7",
-          "#9c755f",
-          "#bab0ac"
-        )[1L:n]
+        colours <- palette_10[1L:n]
         
       } else if (n <= 20L) {
-        colours <- c(
-          "#4e79a7",
-          "#a0cbe8",
-          "#f28e2b",
-          "#ffbe7d",
-          "#59a14f",
-          "#8cd17d",
-          "#b6992d",
-          "#f1ce63",
-          "#499894",
-          "#86bcb6",
-          "#e15759",
-          "#ff9d9a",
-          "#79706e",
-          "#bab0ac",
-          "#d37295",
-          "#fabfd2",
-          "#b07aa1",
-          "#d4a6c8",
-          "#9d7660",
-          "#d7b5a6"
-        )[1L:n]
+        colours <- palette_20[1L:n]
         
       } else {
         ..error(paste0(
@@ -240,42 +248,10 @@
       # This is to avoid confusion with other gradients that may be used in the
       # plot.
       if (n <= 10L) {
-        colours <- c(
-          "#e15759",
-          "#76b7b2",
-          "#59a14f",
-          "#edc948",
-          "#b07aa1",
-          "#ff9da7",
-          "#9c755f",
-          "#bab0ac",
-          "#4e79a7",
-          "#f28e2b"
-        )[1L:n]
+        colours <- rev(palette_10)[1L:n]
         
       } else if (n <= 20L) {
-        colours <- c(
-          "#59a14f",
-          "#8cd17d",
-          "#b6992d",
-          "#f1ce63",
-          "#499894",
-          "#86bcb6",
-          "#e15759",
-          "#ff9d9a",
-          "#79706e",
-          "#bab0ac",
-          "#d37295",
-          "#fabfd2",
-          "#b07aa1",
-          "#d4a6c8",
-          "#9d7660",
-          "#d7b5a6",
-          "#4e79a7",
-          "#a0cbe8",
-          "#f28e2b",
-          "#ffbe7d"
-        )[1L:n]
+        colours <- rev(palette_20)[1L:n]
         
       } else {
         ..error(paste0(
@@ -287,28 +263,22 @@
   } else if (palette_type == "sequential") {
     if (!use_alternative) {
       # A palette with the same hue (blue) as the first color of the qualitative
-      # palette. Based on an advanced single-hue sequential palette created
-      # using colorspace::hcl_wizard, with Hue 1=245, Chroma 1 = 15, Max Chroma
-      # = 75, Lumin. 1 = 20, Lumin 2 = 98, Power 1 = 0.8, Power 2 = 1.4, without
-      # color correction.
+      # palette. Based on the blue part of colorspace::diverge_hcl(n = 21, h =
+      # c(245, 35), c = c(130, 85), l = c(65, 80), power = c(1, 1.3), gamma =
+      # NULL, fixup = TRUE)
       colours <- c(
-        "#233143", "#243950", "#26415D", "#27496A", "#285177", "#295A85",
-        "#296393", "#2A6CA2", "#2A76B1", "#297FC0", "#2D89CF", "#4D93D2",
-        "#649CD6", "#79A7DA", "#8CB1DF", "#9FBDE3", "#B2C9E8", "#C6D6ED",
-        "#DCE5F2", "#DCDCDC"
+        "#C6C6C6", "#BCC5D2", "#B0C3DD", "#A3C0E7", "#93BDF0", "#81BAF8",
+        "#6AB7FF", "#49B3FF", "#00B0FF", "#00ACFF", "#00A8FF"
       )
+      
     } else {
       # Alternative reddish colour scheme that avoids the use of blues and
-      # orange tones that may have been used as a primary palette. Based on an
-      # advanced single-hue sequential palette created using
-      # colorspace::hcl_wizard, with Hue=12, Chroma 1 = 40, Max Chroma = 120,
-      # Lumin. 1 = 20, Lumin 2 = 98, Power 1 = 0.8, Power 2 = 1.4, without color
-      # correction.
+      # orange tones that may have been used as a primary palette. Based on the
+      # red part of colorspace::diverge_hcl(n = 21, h = c(185, 10), c = c(130,
+      # 85), l = c(60, 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
       colours <- c(
-        "#581B1C", "#661F1F", "#732323", "#812727", "#8F2B2C", "#9E2F30",
-        "#AC3434", "#BC3839", "#CC3D3E", "#DC4243", "#E25354", "#E76364",
-        "#EC7374", "#F18383", "#F59394", "#F8A4A4", "#FBB5B6", "#FCC8C8",
-        "#FCDDDD", "#DCDCDC"
+        "#C6C6C6", "#D4BFBF", "#DFB6B7", "#E8ACAE", "#EFA1A4", "#F5969A",
+        "#F98A8F", "#FC7E84", "#FE7178", "#FE636C", "#FE535E"
       )
     }
 
@@ -316,71 +286,62 @@
     
   } else if (palette_type == "divergent") {
     if (!use_alternative) {
-      # A palette with the same hues (blue and orange) as the first two colors of
-      # the qualitative palette. Based on a combination of two advanced single-hue
-      # sequential palettes created using colorspace::hcl_wizard:
-      #
-      # Blues: Hue 1=245, Chroma 1 = 15, Max Chroma = 75, Lumin. 1 = 20, Lumin 2 =
-      # 98, Power 1 = 0.8, Power 2 = 1.4, without color correction. Oranges: Hue
-      # 1=36, Chroma 1 = 20, Max Chroma = 100, Lumin. 1 = 35, Lumin 2 = 98, Power
-      # 1 = 0.8, Power 2 = 1.4, without color correction, in reverse order.
+      # A palette with the same hues (blue and orange) as the first two colors
+      # of the qualitative palette.
       if (!diverge_to_white) {
-        # Centre colour is white.
+        # Bright center.
+        #
+        # colorspace::diverge_hcl(n = 21, h = c(245, 35), c = c(130, 85), l =
+        # c(65, 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
         colours <- c(
-          "#233143", "#243950", "#26415D", "#27496A", "#285177", "#295A85",
-          "#296393", "#2A6CA2", "#2A76B1", "#297FC0", "#2D89CF", "#4D93D2",
-          "#649CD6", "#79A7DA", "#8CB1DF", "#9FBDE3", "#B2C9E8", "#C6D6ED", "#DCE5F2",
-          "#DCDCDC",
-          "#FAE4DA", "#F9D4C3", "#F8C6AE", "#F5B999", "#F2AD85", "#EFA170",
-          "#EB965A", "#E78C40", "#E28118", "#D67B1E", "#C87528", "#BC702F",
-          "#AF6A34", "#A36537", "#96603A", "#8A5B3C", "#7E563E", "#725240", "#664D41"
+          "#00A8FF", "#00ACFF", "#00B0FF", "#49B3FF", "#6AB7FF", "#81BAF8",
+          "#93BDF0", "#A3C0E7", "#B0C3DD", "#BCC5D2",
+          "#C6C6C6",
+          "#D2C1BA", "#DBBBAC", "#E3B49D", "#E9AD8D", "#EEA57C", "#F29D68",
+          "#F49650", "#F68E2F", "#F78600", "#F77D00"
         )
+        
       } else {
-        # Centre colour is dark.
+        # Dark centre.
+        #
+        # colorspace::diverge_hcl(n = 21, h = c(245, 35), c = c(130, 85), l =
+        # c(65, 30), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
         colours <- c(
-          "#FAE4DA", "#F9D4C3", "#F8C6AE", "#F5B999", "#F2AD85", "#EFA170",
-          "#EB965A", "#E78C40", "#E28118", "#D67B1E", "#C87528", "#BC702F",
-          "#AF6A34", "#A36537", "#96603A", "#8A5B3C", "#7E563E", "#725240", "#664D41",
-          "#050505",
-          "#233143", "#243950", "#26415D", "#27496A", "#285177", "#295A85",
-          "#296393", "#2A6CA2", "#2A76B1", "#297FC0", "#2D89CF", "#4D93D2",
-          "#649CD6", "#79A7DA", "#8CB1DF", "#9FBDE3", "#B2C9E8", "#C6D6ED", "#DCE5F2"
+          "#00A8FF", "#009BFF", "#008FEF", "#0083D8", "#0078C1", "#056DAB",
+          "#266495", "#345B81", "#3C526D", "#424C59",
+          "#474747",
+          "#584740", "#694A39", "#794E32", "#8A522A", "#9A581E", "#AC5E07",
+          "#BE6500", "#D06D00", "#E37500" ,"#F77D00"
         )
       }
+      
     } else {
       # A palette based on the same hues the first two colours of the
-      # alternative qualitative palette. Based on a combination of two advanced
-      # single-hue sequential palettes created using colorspace::hcl_wizard:
-      #
-      # Reds: Hue=12, Chroma 1 = 40, Max Chroma = 120, Lumin. 1 = 20, Lumin 2 =
-      # 98, Power 1 = 0.8, Power 2 = 1.4, without color correction. Cyan:
-      # Hue=185, Chroma 1 = 0, Max Chroma = 55, Lumin. 1 = 40, Lumin 2 = 98,
-      # Power 1 = 1.0, Power 2 = 1.4, without color correction.
+      # alternative qualitative palette.
       if (!diverge_to_white) {
-        # Centre colour is white.
+        # Bright centre.
+        #
+        # colorspace::diverge_hcl(n = 21, h = c(185, 10), c = c(130, 85), l =
+        # c(60, 80), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
         colours <- c(
-          "#581B1C", "#661F1F", "#732323", "#812727", "#8F2B2C", "#9E2F30",
-          "#AC3434", "#BC3839", "#CC3D3E", "#DC4243", "#E25354", "#E76364",
-          "#EC7374", "#F18383", "#F59394", "#F8A4A4", "#FBB5B6", "#FCC8C8", "#FCDDDD",
-          "#DCDCDC",
-          "#E8F3F2", "#D7ECEA", "#C5E6E3", "#B2DFDC", "#9FD9D4", "#8AD3CD",
-          "#72CCC6", "#56C6BF", "#2DC0B8", "#16B7B0", "#30ADA6", "#3EA39D",
-          "#489994", "#4F8F8B", "#548582", "#587B79", "#5B7270", "#5D6867", "#5E5E5E"
+          "#00BCAB", "#00BFAF", "#00C2B4", "#00C4B7", "#00C6BB", "#00C8BE",
+          "#40C9C1", "#74C9C3", "#96C9C5", "#B0C9C6",
+          "#C6C6C6",
+          "#D4BFBF", "#DFB6B7", "#E8ACAE", "#EFA1A4", "#F5969A", "#F98A8F",
+          "#FC7E84", "#FE7178", "#FE636C", "#FE535E"
         )
+        
       } else {
-        # Centre colour is black.
+        # Dark centre.
+        #
+        # colorspace::diverge_hcl(n = 21, h = c(185, 10), c = c(130, 85), l =
+        # c(60, 30), power = c(1, 1.3), gamma = NULL, fixup = TRUE)
         colours <- c(
-          rev(c(
-            "#E8F3F2", "#D7ECEA", "#C5E6E3", "#B2DFDC", "#9FD9D4", "#8AD3CD",
-            "#72CCC6", "#56C6BF", "#2DC0B8", "#16B7B0", "#30ADA6", "#3EA39D",
-            "#489994", "#4F8F8B", "#548582", "#587B79", "#5B7270", "#5D6867", "#5E5E5E"
-          )),
-          "#050505",
-          rev(c(
-            "#581B1C", "#661F1F", "#732323", "#812727", "#8F2B2C", "#9E2F30",
-            "#AC3434", "#BC3839", "#CC3D3E", "#DC4243", "#E25354", "#E76364",
-            "#EC7374", "#F18383", "#F59394", "#F8A4A4", "#FBB5B6", "#FCC8C8", "#FCDDDD"
-          ))
+          "#00BCAB", "#00AE9F", "#00A093", "#009387", "#00877C", "#007A71",
+          "#006F67", "#00645E", "#065955", "#334F4D",
+          "#474747",
+          "#5A4546", "#6D4446", "#7E4547", "#904649", "#A1474C", "#B3494F",
+          "#C54B52", "#D74D56", "#EB505A", "#FE535E"
         )
       }
     }